requirements.txt
# Production dependencies
discord.py==2.6.4
python-dotenv==1.2.1

# Transitive dependencies
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
yarl==1.22.0

# Development dependencies
black==26.1.0
flake8==7.3.0
mypy==1.19.1
mypy_extensions==1.1.0
click==8.3.1
colorama==0.4.6
librt==0.7.8
mccabe==0.7.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.4.0
typing_extensions==4.15.0
